% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.tsa_hr}
\alias{print.tsa_hr}
\title{Print a tsa_hr object}
\usage{
\method{print}{tsa_hr}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"tsa_hr"}.}

\item{...}{Currently unused.}
}
\value{
An object of class \code{"tsa_hr"}, namely the same object
  \code{x} supplied to the method, returned invisibly. The method
  prints a concise summary of the Trial Sequential Analysis results
  and returns the original object unchanged.
}
\description{
Print a tsa_hr object
}
