% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{tsahr-package}
\alias{tsahr}
\alias{tsahr-package}
\title{tsahr: Trial Sequential Analysis for Meta-Analyses of Hazard Ratios}
\description{
Performs Trial Sequential Analysis (TSA) for meta-analyses of time-to-event outcomes reported as hazard ratios. Implements the Schoenfeld required-events sample-size formula generalised for unequal allocation, the Diversity (D-squared) heterogeneity adjustment of Wetterslev et al. (2009), and O'Brien-Fleming-type alpha- and beta-spending trial sequential monitoring boundaries computed at the inverse-variance information based on observed study-level log-HR standard errors (not merely event counts), via a compiled (C++) recursive numerical integration engine ported from the R package 'RTSA' (Soerensen, Olsen, Lange and Gluud; an R implementation of the Trial Sequential Analysis software of the Copenhagen Trial Unit, \url{https://ctu.dk/tools}), with no fixed software-imposed limit on the number of looks (subject to available computational resources). Produces a cumulative Z-curve plot with efficacy, futility, and conventional significance boundaries. Methodology follows Miladinovic et al. (2013) \doi{10.1016/j.jclinepi.2012.11.007} and Wetterslev et al. (2009) \doi{10.1186/1471-2288-9-86}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tarak-dhaouadi/tsahr}
  \item Report bugs at \url{https://github.com/tarak-dhaouadi/tsahr/issues}
}

}
\author{
\strong{Maintainer}: Tarak Dhaouadi \email{dhaouaditarak@yahoo.fr}

Authors:
\itemize{
  \item Tarak Dhaouadi \email{dhaouaditarak@yahoo.fr}
}

Other contributors:
\itemize{
  \item Anne Lyngholm Soerensen (author of the RTSA package, from which the boundary engine is ported) [contributor, copyright holder]
  \item Markus Harboe Olsen (author of the RTSA package, from which the boundary engine is ported) [contributor, copyright holder]
  \item Theis Lange (author of the RTSA package, from which the boundary engine is ported) [contributor, copyright holder]
  \item Christian Gluud (author of the RTSA package, from which the boundary engine is ported) [contributor, copyright holder]
}

}
