\name{summary.tsa_hr}
\alias{summary.tsa_hr}
\title{Summarise a tsa_hr object}
\usage{
\method{summary}{tsa_hr}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"tsa_hr"}.}

\item{...}{Currently unused.}
}
\value{
The underlying summary data.frame (invisibly printed).
}
\description{
Summarise a tsa_hr object
}
